backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1
graphql-core==3.2.3

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Request handling
python-multipart==0.0.6

# Hedera Blockchain SDK - OFFICIAL
hedera-sdk-py==2.50.0
web3==6.11.3

# HTTP clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Email functionality
aiosmtplib==3.0.1
jinja2==3.1.2

# Environment and configuration
python-dotenv==1.0.0

# CORS middleware
# Note: FastAPI has built-in CORS, fastapi-cors is not needed
# Removed: fastapi-cors==0.0.6

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# JSON handling
orjson==3.9.10

# Logging and monitoring
structlog==23.2.0

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing async endpoints

# Development tools (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production WSGI server (alternative to uvicorn)
gunicorn==21.2.0

# Additional utilities
click==8.1.7  # For CLI commands
rich==13.7.0  # For better console output
typer==0.9.0  # For CLI interface

# Redis for nonce caching
redis==5.0.1

# Additional crypto utilities
eth-account==0.9.0
eth-utils==2.3.0